+++ aeronet/server/test/lifecycle_test.cpp
@@ -3,6 +3,7 @@
 #include <gtest/gtest.h>
 
 #include <chrono>
+#include <stdexcept>
 #include <thread>
 #include <utility>